:wastebasket:Delete and :truck:moved files *Deleted unused data /Items *Deployed :rocket *Added gh-pages dependency
