+++ .gitignore
@@ -1,0 +1,21 @@
+# dependencies
+/node_modules
+# production
+/build
+.idea/*
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
